Move global markdown object cache into PackageGraph
